requirements.txt
contourpy==1.1.1
cycler==0.12.1
fonttools==4.43.1
kiwisolver==1.4.5
llvmlite==0.41.1
matplotlib==3.8.0
numba==0.58.1
numpy==1.26.1
packaging==23.2
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
